move action

+++ src/point.rs
@@ -2,12 +2,12 @@ use std::ops::Add;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Point {
-    pub x: usize,
-    pub y: usize,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Point {
-    pub fn new(x: usize, y: usize) -> Point {
+    pub fn new(x: i32, y: i32) -> Point {
         Point { x, y }
     }
 }
